Struct lspower::Client

pub struct Client { /* fields omitted */ }

Handle for communicating with the language client.

This type provides a very cheap implementation of Clone so API consumers can cheaply clone and pass it around as needed.

Implementations

impl Client

pub async fn log_message<M: Display>(&self, typ: MessageType, message: M)

Notifies the client to log a particular message.

This corresponds to the window/logMessage notification.

pub async fn show_message<M: Display>(&self, typ: MessageType, message: M)

Notifies the client to display a particular message in the user interface.

This corresponds to the window/showMessage notification.

pub async fn show_message_request<M: Display>(
&self,
typ: MessageType,
message: M,
actions: Option<Vec<MessageActionItem>>
) -> Result<Option<MessageActionItem>>

Requests the client to display a particular message in the user interface.

Unlike the show_message notification, this request can also pass a list of actions and wait for an answer from the client.

This corresponds to the window/showMessageRequest request.

pub async fn telemetry_event<S: Serialize>(&self, data: S)

Notifies the client to log a telemetry event.

This corresponds to the telemetry/event notification.

pub async fn register_capability(
&self,
registrations: Vec<Registration>
) -> Result<()>

Registers a new capability with the client.

This corresponds to the client/registerCapability request.

Initialization

If the request is sent to client before the server has been initialized, this will immediately return Err with JSON-RPC error code -32002 (read more).

pub async fn unregister_capability(
&self,
unregisterations: Vec<Unregistration>
) -> Result<()>

Unregisters a capability with the client.

This corresponds to the client/unregisterCapability request.

Initialization

If the request is sent to client before the server has been initialized, this will immediately return Err with JSON-RPC error code -32002 (read more).

pub async fn workspace_folders(&self) -> Result<Option<Vec<WorkspaceFolder>>>

Fetches the current open list of workspace folders.

Returns None if only a single file is open in the tool. Returns an empty Vec if a workspace is open but no folders are configured.

This corresponds to the workspace/workspaceFolders request.

Initialization

If the request is sent to client before the server has been initialized, this will immediately return Err with JSON-RPC error code -32002 (read more).

Compatibility

This request was introduced in specification version 3.6.0.

pub async fn configuration(
&self,
items: Vec<ConfigurationItem>
) -> Result<Vec<Value>>

Fetches configuration settings from the client.

The request can fetch several configuration settings in one roundtrip. The order of the returned configuration settings correspond to the order of the passed ConfigurationItems (e.g. the first item in the response is the result for the first configuration item in the params).

This corresponds to the workspace/configuration request.

Initialization

If the request is sent to client before the server has been initialized, this will immediately return Err with JSON-RPC error code -32002 (read more).

Compatibility

This request was introduced in specification version 3.6.0.

pub async fn apply_edit(
&self,
edit: WorkspaceEdit,
label: Option<String>
) -> Result<ApplyWorkspaceEditResponse>

Requests a workspace resource be edited on the client side and returns whether the edit was applied.

This corresponds to the workspace/applyEdit request.

Initialization

If the request is sent to client before the server has been initialized, this will immediately return Err with JSON-RPC error code -32002 (read more).

pub async fn publish_diagnostics(
&self,
uri: Url,
diags: Vec<Diagnostic>,
version: Option<i32>
)

Submits validation diagnostics for an open file with the given URI.

This corresponds to the textDocument/publishDiagnostics notification.

Initialization

This notification will only be sent if the server is initialized.

pub async fn send_custom_notification<N>(&self, params: N::Params) where
N: Notification,

Sends a custom notification to the client.

Initialization

This notification will only be sent if the server is initialized.