Crate lspower

Language Server Protocol (LSP) server abstraction for Tower.

Re-exports

pub extern crate lsp;

Modules

jsonrpc

A subset of JSON-RPC types used by the Language Server Protocol.

Structs

CancellationToken

A token which listens for a cancellation signal from a TokenCanceller.

Client

Handle for communicating with the language client.

ExitedError

Error that occurs when attempting to call the language server after it has already exited.

LspService

Service abstraction for the Language Server Protocol.

MessageStream

Stream of messages produced by the language server.

Server

Server for processing requests and responses on standard I/O or TCP.

TokenCanceller

A structure used to construct and cancel CancellationToken.

Traits

LanguageServer

Trait implemented by language server backends.

Attribute Macros

async_trait